requirements.txt
setuptools~=54.0.0
pymongo~=3.11.3
Scrapy~=2.4.1
itemadapter~=0.2.0
python-dotenv~=0.15.0
goodreads~=0.0.0
Flask~=1.1.2
MyApplication~=0.1.0